+++ src/const/blob-types.const.js
@@ -1,0 +1,8 @@
+export const BLOB_TYPES = Object.freeze({
+  assetDescription: 'asset_description',
+  saleOverview: 'fund_overview',
+  saleDocument: 'fund_document',
+  kycCorporate: 'alpha',
+  kycGeneral: 'kyc_form',
+  bravo: 'bravo'
+})